Renamed namespace from postgres_models[.tables] -> postgres_database[.models] Removed all suffix table from models
