+++ packages/postgres-database/alembic.ini
@@ -35,7 +35,7 @@ timezone = UTC
 # are written from script.py.mako
 # output_encoding = utf-8
 
-# url for testing. See packages/postgres_models/tests/docker-compose.yml
+# url for testing. See packages/postgres-database/tests/docker-compose.yml
 sqlalchemy.url = None
 #sqlalchemy.url = postgresql+psycopg2://test:test@localhost:5432/test
 #sqlalchemy.url = postgresql+psycopg2://simcore:simcore@localhost:30001/simcoredb
