+++ packages/postgres-database/migration/env.py
@@ -2,7 +2,7 @@
 from logging.config import fileConfig
 
 from alembic import context
-from simcore_postgres_models.settings import target_metadatas, sqlalchemy_url
+from simcore_postgres_database.settings import target_metadatas, sqlalchemy_url
 from sqlalchemy import engine_from_config, pool
 
 # this is the Alembic Config object, which provides
@@ -36,10 +36,8 @@ def run_migrations_offline():
 
     """
     # url = config.get_main_option("sqlalchemy.url")
-
-
     context.configure(
-        url=url, target_metadata=target_metadata, literal_binds=True
+        url=sqlalchemy_url, target_metadata=target_metadata, literal_binds=True
     )
 
     with context.begin_transaction():
@@ -53,8 +51,11 @@ def run_migrations_online():
     and associate a connection with the context.
 
     """
+    config_dict = config.get_section(config.config_ini_section)
+    config_dict['sqlalchemy.url'] = sqlalchemy_url
+
     connectable = engine_from_config(
-        config.get_section(config.config_ini_section),
+        config_dict,
         prefix="sqlalchemy.",
         poolclass=pool.NullPool,
     )
