+++ packages/postgres-database/src/simcore_postgres_database/__init__.py
@@ -1,0 +1,8 @@
+from . import storage_models, webserver_models
+from .models.base import metadata
+
+__all__ = [
+    'metadata',
+    'webserver_models',
+    'storage_models'
+]
