+++ packages/postgres-database/src/simcore_postgres_database/storage_models.py
@@ -3,9 +3,9 @@
     - Facade to direct access to models in the database by
     the storage service
 """
-from .tables.base import metadata
-from .tables.file_meta_data_table import file_meta_data
-from .tables.tokens_table import tokens
+from .models.base import metadata
+from .models.file_meta_data import file_meta_data
+from .models.tokens import tokens
 
 __all__ = [
     "tokens",
