+++ packages/postgres-database/src/simcore_postgres_database/webserver_models.py
@@ -1,0 +1,23 @@
+""" Facade for webserver service
+
+    Facade to direct access to models in the database by
+    the  webserver service
+
+"""
+from .models.comp_pipeline import comp_pipeline
+from .models.comp_tasks import comp_tasks
+from .models.confirmations import ConfirmationAction, confirmations
+from .models.projects import ProjectType, projects
+from .models.tokens import tokens
+from .models.user_to_projects import user_to_projects
+from .models.users import UserRole, UserStatus, users
+
+__all__ = [
+    "users", "UserRole", "UserStatus",
+    "projects", "ProjectType",
+    "user_to_projects",
+    "confirmations", "ConfirmationAction",
+    "tokens",
+    "comp_tasks",
+    "comp_pipeline"
+]
