+++ packages/postgres-database/tests/docker-compose.yml
@@ -6,8 +6,10 @@ services:
       - POSTGRES_USER=test
       - POSTGRES_PASSWORD=test
       - POSTGRES_DB=test
+      - POSTGRES_HOST=localhost
+      - POSTGRES_PORT=5432
     volumes:
-      - test_postgres_models:/var/lib/postgresql/data
+      - test_postgres_database_volume:/var/lib/postgresql/data
     ports:
       - "5432:5432"
   adminer:
@@ -17,4 +19,4 @@ services:
     depends_on:
       - postgres
 volumes:
-  test_postgres_models:
+  test_postgres_database_volume:
