--- packages/postgres-models/src/simcore_postgres_models/webserver_tables.py
@@ -1,23 +1,0 @@
-""" Facade for webserver service
-
-    Facade to direct access to models in the database by
-    the  webserver service
-
-"""
-from .tables.comp_pipeline_table import comp_pipeline
-from .tables.comp_tasks_table import comp_tasks
-from .tables.confirmations_table import ConfirmationAction, confirmations
-from .tables.projects_table import ProjectType, projects
-from .tables.tokens_table import tokens
-from .tables.user_to_projects_table import user_to_projects
-from .tables.users_table import UserRole, UserStatus, users
-
-__all__ = [
-    "users", "UserRole", "UserStatus",
-    "projects", "ProjectType",
-    "user_to_projects",
-    "confirmations", "ConfirmationAction",
-    "tokens",
-    "comp_tasks",
-    "comp_pipeline"
-]
