+++ packages/simcore-sdk/src/simcore_sdk/models/DEPRECATED
@@ -1,1 +1,1 @@
-To get models, use instead  ``import simcore_postgres_models``
+To get models, use instead  ``import simcore_postgres_database``
