+++ services/storage/src/simcore_service_storage/models.py
@@ -5,7 +5,7 @@ from typing import Tuple
 
 import attr
 
-from simcore_postgres_models.storage_tables import (file_meta_data, metadata,
+from simcore_postgres_database.storage_tables import (file_meta_data, metadata,
                                                     tokens)
 
 from .settings import DATCORE_STR, SIMCORE_S3_ID, SIMCORE_S3_STR
