+++ services/web/server/src/simcore_service_webserver/security_roles.py
@@ -4,7 +4,7 @@
     The access model is created upon setting up of the security subsystem
 """
 
-from simcore_postgres_models.tables.users_table import UserRole
+from simcore_postgres_database.models.users_table import UserRole
 
 
 # A role defines a set of operations that the user *can* perform
